# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Beer Mile multi-tenant event tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  super_admin: 'super_admin',
  host: 'host',
  runner: 'runner',
} as const;

/**
 * @nullable
 */
export type UserGender = typeof UserGender[keyof typeof UserGender] | null;


export const UserGender = {
  male: 'male',
  female: 'female',
  non_binary: 'non_binary',
  prefer_not_to_say: 'prefer_not_to_say',
} as const;

export interface User {
  id: string;
  email: string;
  role: UserRole;
  /** @nullable */
  gender?: UserGender;
  /** @nullable */
  birthdate?: string | null;
  /** @nullable */
  preferredName?: string | null;
  authProvider?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface UserProvision {
  preferredName?: string;
}

export type UserProfileUpdateGender = typeof UserProfileUpdateGender[keyof typeof UserProfileUpdateGender];


export const UserProfileUpdateGender = {
  male: 'male',
  female: 'female',
  non_binary: 'non_binary',
  prefer_not_to_say: 'prefer_not_to_say',
} as const;

export interface UserProfileUpdate {
  gender?: UserProfileUpdateGender;
  birthdate?: string;
  preferredName?: string;
}

export interface Tenant {
  id: number;
  userId: string;
  organizationName: string;
  createdAt: string;
}

export interface TenantInput {
  /** @minLength 1 */
  organizationName: string;
}

export interface TenantWithUser {
  id: number;
  userId: string;
  organizationName: string;
  createdAt: string;
  /** @nullable */
  userEmail?: string | null;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  draft: 'draft',
  open: 'open',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Prize {
  position: number;
  description: string;
}

export interface Event {
  id: number;
  tenantId: number;
  title: string;
  eventCode: string;
  eventDate: string;
  /** @nullable */
  locationLat?: number | null;
  /** @nullable */
  locationLng?: number | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  beerType?: string | null;
  /** @nullable */
  entryFee?: string | null;
  /** @nullable */
  paymentInstructions?: string | null;
  /** @nullable */
  prizesJson?: Prize[] | null;
  status: EventStatus;
  createdAt: string;
  updatedAt?: string;
}

export type EventInputStatus = typeof EventInputStatus[keyof typeof EventInputStatus];


export const EventInputStatus = {
  draft: 'draft',
  open: 'open',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventInput {
  /** @minLength 1 */
  title: string;
  eventDate: string;
  locationLat?: number;
  locationLng?: number;
  locationName?: string;
  beerType?: string;
  entryFee?: string;
  paymentInstructions?: string;
  prizesJson?: Prize[];
  status?: EventInputStatus;
}

export type EventUpdateStatus = typeof EventUpdateStatus[keyof typeof EventUpdateStatus];


export const EventUpdateStatus = {
  draft: 'draft',
  open: 'open',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  eventDate?: string;
  locationLat?: number;
  locationLng?: number;
  locationName?: string;
  beerType?: string;
  entryFee?: string;
  paymentInstructions?: string;
  prizesJson?: Prize[];
  status?: EventUpdateStatus;
}

export interface EventCodeLookup {
  eventCode: string;
}

export type RegistrationPaymentStatus = typeof RegistrationPaymentStatus[keyof typeof RegistrationPaymentStatus];


export const RegistrationPaymentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
} as const;

export interface Registration {
  id: number;
  eventId: number;
  userId: string;
  paymentStatus: RegistrationPaymentStatus;
  /** @nullable */
  tagUid?: string | null;
  /** @nullable */
  runnerToken?: string | null;
  registeredAt: string;
  updatedAt?: string;
}

export type RegistrationWithUserPaymentStatus = typeof RegistrationWithUserPaymentStatus[keyof typeof RegistrationWithUserPaymentStatus];


export const RegistrationWithUserPaymentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
} as const;

export interface RegistrationWithUser {
  id: number;
  eventId: number;
  userId: string;
  paymentStatus: RegistrationWithUserPaymentStatus;
  /** @nullable */
  tagUid?: string | null;
  /** @nullable */
  runnerToken?: string | null;
  registeredAt: string;
  updatedAt?: string;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  preferredName?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  birthdate?: string | null;
}

export type RegistrationUpdatePaymentStatus = typeof RegistrationUpdatePaymentStatus[keyof typeof RegistrationUpdatePaymentStatus];


export const RegistrationUpdatePaymentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
} as const;

export interface RegistrationUpdate {
  paymentStatus?: RegistrationUpdatePaymentStatus;
  tagUid?: string;
  runnerToken?: string;
}

export interface LapLogEntry {
  clientEventLogId: string;
  registrationId: number;
  lapNumber: number;
  /** @nullable */
  splitTimeMs?: number | null;
  elapsedMs: number;
  pourConfirmed: boolean;
  /** @nullable */
  deviceMonotonicTimestamp?: number | null;
}

export interface LapLogBatch {
  records: LapLogEntry[];
}

export interface LapLogIngestResult {
  processedIds: string[];
}

export interface LapSplit {
  lapNumber: number;
  elapsedMs: number;
  /** @nullable */
  splitTimeMs?: number | null;
  pourConfirmed?: boolean;
}

export interface LeaderboardEntry {
  registrationId: number;
  userId: string;
  /** @nullable */
  preferredName?: string | null;
  /** @nullable */
  gender?: string | null;
  totalElapsedMs: number;
  lapCount: number;
  finished?: boolean;
  laps: LapSplit[];
}

export type AdminAnalyticsEventsByStatus = {[key: string]: number};

export interface AdminAnalytics {
  totalEvents: number;
  totalRunners: number;
  totalTenants: number;
  totalFinishers: number;
  /** @nullable */
  avgLapTimeMs: number | null;
  eventsByStatus?: AdminAnalyticsEventsByStatus;
}

export interface PlatformSettings {
  id: number;
  /** @nullable */
  appStoreIosUrl?: string | null;
  /** @nullable */
  playStoreAndroidUrl?: string | null;
  updatedAt?: string;
}

export interface PlatformSettingsUpdate {
  appStoreIosUrl?: string;
  playStoreAndroidUrl?: string;
}