# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beer Mile multi-tenant event tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events (public browsing + host's own events)
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/join:
    post:
      operationId: joinEventByCode
      tags: [events]
      summary: Find an event by private event code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCodeLookup"
      responses:
        "200":
          description: Found event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (host or super_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (host or super_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
  /events/{id}/registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List registrations for an event (host or super_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of registrations with user info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationWithUser"
        "403":
          description: Forbidden
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Register current user for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "409":
          description: Already registered
  /registrations/{id}:
    patch:
      operationId: updateRegistration
      tags: [registrations]
      summary: Update registration (payment status, tag UID, runner token)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationUpdate"
      responses:
        "200":
          description: Updated registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /events/{id}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get leaderboard for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "404":
          description: Not found
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, tenantId, title, eventCode, eventDate, status, createdAt]
      properties:
        id:
          type: number
        tenantId:
          type: number
        title:
          type: string
        eventCode:
          type: string
        eventDate:
          type: string
        locationLat:
          type: ["number", "null"]
        locationLng:
          type: ["number", "null"]
        locationName:
          type: ["string", "null"]
        beerType:
          type: ["string", "null"]
        entryFee:
          type: ["string", "null"]
        paymentInstructions:
          type: ["string", "null"]
        prizesJson:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/Prize"
        status:
          type: string
          enum: [draft, open, active, completed, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Prize:
      type: object
      required: [position, description]
      properties:
        position:
          type: number
        description:
          type: string
    EventInput:
      type: object
      required: [title, eventDate]
      properties:
        title:
          type: string
          minLength: 1
        eventDate:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        locationName:
          type: string
        beerType:
          type: string
        entryFee:
          type: string
        paymentInstructions:
          type: string
        prizesJson:
          type: array
          items:
            $ref: "#/components/schemas/Prize"
        status:
          type: string
          enum: [draft, open, active, completed, cancelled]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        eventDate:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        locationName:
          type: string
        beerType:
          type: string
        entryFee:
          type: string
        paymentInstructions:
          type: string
        prizesJson:
          type: array
          items:
            $ref: "#/components/schemas/Prize"
        status:
          type: string
          enum: [draft, open, active, completed, cancelled]
    EventCodeLookup:
      type: object
      required: [eventCode]
      properties:
        eventCode:
          type: string
    Registration:
      type: object
      required: [id, eventId, userId, paymentStatus, registeredAt]
      properties:
        id:
          type: number
        eventId:
          type: number
        userId:
          type: string
        paymentStatus:
          type: string
          enum: [pending, confirmed]
        tagUid:
          type: ["string", "null"]
        runnerToken:
          type: ["string", "null"]
        registeredAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegistrationWithUser:
      type: object
      required: [id, eventId, userId, paymentStatus, registeredAt]
      properties:
        id:
          type: number
        eventId:
          type: number
        userId:
          type: string
        paymentStatus:
          type: string
          enum: [pending, confirmed]
        tagUid:
          type: ["string", "null"]
        runnerToken:
          type: ["string", "null"]
        registeredAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userEmail:
          type: ["string", "null"]
        preferredName:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
    RegistrationUpdate:
      type: object
      properties:
        paymentStatus:
          type: string
          enum: [pending, confirmed]
        tagUid:
          type: string
        runnerToken:
          type: string
...
    LeaderboardEntry:
      type: object
      required: [registrationId, userId, totalElapsedMs, lapCount, laps]
      properties:
        registrationId:
          type: number
        userId:
          type: string
        preferredName:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        totalElapsedMs:
          type: number
        lapCount:
          type: number
        finished:
          type: boolean
        laps:
          type: array
          items:
            $ref: "#/components/schemas/LapSplit"
    LapSplit:
      type: object
      required: [lapNumber, elapsedMs]
      properties:
        lapNumber:
          type: number
        elapsedMs:
          type: number
        splitTimeMs:
          type: ["number", "null"]
        pourConfirmed:
          type: boolean
...