# lib/api-zod/src/generated/types/lapSplit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Beer Mile multi-tenant event tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface LapSplit {
  lapNumber: number;
  elapsedMs: number;
  /** @nullable */
  splitTimeMs?: number | null;
  pourConfirmed?: boolean;
}